+++ src/Components/GlobalComponents/Header/hooks/useHeader.js
@@ -1,0 +1,13 @@
+import { useRef } from "react";
+import { useResize } from "hooks/useResize";
+
+export const useHeader = () => {
+  const closeOffCanvas = useRef(null);
+  const { isPhone } = useResize();
+
+  const handleCloseOffCanvas = () => {
+    closeOffCanvas.current.click();
+  };
+
+  return { closeOffCanvas, isPhone, handleCloseOffCanvas };
+};
